Reduce the number of worker
